requirements.txt
Django==1.8.3
Markdown==2.6.2
PyDispatcher==2.0.5
PyJWT==1.4.0
Pygments==2.0.2
SQLAlchemy==1.0.8
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.11.0
django-postgrespool==0.3.0
django-sslserver==0.15
django-stormpath==1.0.3
django-toolbelt==0.0.1
djangorestframework==3.2.4
geopy==1.11.0
gevent==1.0.2
greenlet==0.4.7
gunicorn==19.3.0
isodate==0.5.4
oauthlib==1.0.3
psycopg2==2.6.1
python-dateutil==2.4.2
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
static3==0.6.1
stormpath==2.1.1
uWSGI==2.0.11.1
wsgiref==0.1.2
